ReferenceObject: Enable move construction
Before, a ReferenceObject<_T> could not be constructed by an rvalue
reference (_T &&). Enable the construction from such references, but
keep the existing possibility to construct from const lvalue
(const _T &). The implementation is not so nice, but it works for our
projects. If you come with a simpler solution, please let me know.